# PERF: DatetimeIndex.union