```
In [2]: dti = pd.date_range('2016-01-01', periods=10**5, freq="s")

In [3]: dti2 = dti.delete(50000)

In [4]: %timeit dti.union(dti2)
1.77 ms ± 7.97 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
1.55 ms ± 9.29 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```